fs: Extract fuse_driver
