+++ src/fs/driver/fuse/Mybuild
@@ -1,12 +1,14 @@
 package embox.fs.driver
 
-module ext2fuse {
+@BuildDepends(third_party.ext2fuse.core)
+module fuse {
 	@IncludePath("$(EXTERNAL_BUILD_DIR)/third_party/fuse/core/install/include")
 	@IncludePath("$(ROOT_DIR)/third-party/fuse/ext2fuse/include")
 	@IncludePath("$(ROOT_DIR)/third-party/fuse/fuse")
-	source "ext2fuse.c"
+	source "fuse_fs_driver.c"
+
+	@IncludeExport(path="fs")
+	source "fuse_driver.h"
 
 	@NoRuntime depends third_party.ext2fuse.core
-	depends embox.fs.fuse.fuse_linux
-	depends embox.fs.driver.fuse
 }
