+++ src/fs/fuse/fuse_module_linux.h
@@ -26,18 +26,5 @@ extern int fuse_module_umount(struct fuse_module *fm);
 	ARRAY_SPREAD_DECLARE(const struct fuse_module *, fuse_module_repo); \
 	ARRAY_SPREAD_ADD( fuse_module_repo, &fm##__LINE__)
 
-struct fuse_lowlevel_ops;
-struct task;
-struct thread;
-
-struct fuse_sb_priv_data {
-	struct fuse_lowlevel_ops *fuse_lowlevel_ops;
-	struct task *fuse_task;
-	struct thread *stub_thread;
-};
-
-extern const struct super_block_operations fuse_sbops;
-extern const struct inode_operations fuse_iops;
-extern const struct file_operations fuse_fops;
 
 #endif /* SRC_FS_FUSE_FUSE_MODULE_H_ */
